Update README.md
With notifies :restart in the call to the LWRP, chef-client seems to restart the stunnel processes on every run, rather than when there are changes. :reload seems to at least avoid the few second delay the (ubuntu) init script has on restart. Furthermore it seems like if I change the call to the LWRP (even with :reload), chef-client restarts the daemon anyway. But that's preferable since there's an actual change.